added country autocomplete, book lang autocomplete, changed address autocomplete
Changed only the staging file; the production script remains the same. Added book language detection.
