+++ components/disableBtn.js
@@ -1,0 +1,15 @@
+export function disableBtn(fill) {
+  fill = false;
+  //next button style
+  $('[data-form="next-btn"]').css({
+    opacity: "0.4",
+    "pointer-events": "none",
+  });
+  $('[data-form="next-btn"]').addClass("disabled");
+  //submit btn style
+  $('[data-form="submit-btn"]').css({
+    opacity: "0.4",
+    "pointer-events": "none",
+  });
+  $('[data-form="submit-btn"]').addClass("disabled");
+}
