+++ components/enableBtn.js
@@ -1,0 +1,15 @@
+export function enableBtn(fill) {
+  fill = true;
+  //next button style
+  $('[data-form="next-btn"]').css({
+    "pointer-events": "auto",
+    opacity: "1",
+  });
+  $('[data-form="next-btn"]').removeClass("disabled");
+  //submit btn style
+  $('[data-form="submit-btn"]').css({
+    "pointer-events": "auto",
+    opacity: "1",
+  });
+  $('[data-form="submit-btn"]').removeClass("disabled");
+}
