+++ staging-multi-step.js
@@ -1,6 +1,10 @@
 //22-2-23 Update Push
-// AIzaSyCRMOibgbPFCnvJ1IZjVTIVgYPRE7pk0rE - local
-// AIzaSyBacJ90x1fjGipjArXGoRhC4eKijd9mjdU - staging/production
+
+import { changeImage } from "./components/changeImage";
+import { validation } from "./components/validation.js";
+import "./components/location";
+import "./components/countryNumberAutocomplete";
+import "./components/submitForm";
 
 let x = 0;
 let curStep = 0;
@@ -45,37 +49,33 @@ let selArr = [];
 let selString = [];
 let emptyInput = 0;
 let searchQ = [];
-let domainAllowed = true;
-let dom = [];
 let image_changed = false;
 let is_boy = true;
 
 // added new variables
-let autocompleteCity;
-let autocompleteStr;
-let countryInputField = document.querySelector("#Country");
-let streetInputField = document.querySelector("#Street");
-let cityInputField = document.querySelector("#City");
-let zipCode = document.getElementById("ZipCode");
-const environment = document.querySelector("#Environment");
 const host = urlFormly.host;
 const port = urlFormly.port; // if live server is used, then the port is not empty
 const bookLang = document.getElementById("BookLanguage");
 const dateInput = document.getElementById("HeroDOB");
-const phoneInputField = document.querySelector("#Phone");
 
 const detectBookLang = () => {
   const splittedHost = host.split(".");
-  if (splittedHost[0] === "www") {
-    bookLang.value = "en";
-  }
-  //   else if (splittedHost[0] === "uk") {
-  //     bookLang.value = "ua";
-  //   }
-  else {
-    bookLang.value = splittedHost[0];
+  const detectedLanguage = splittedHost[0];
+  for (let i = 0; i < bookLang.options.length; i++) {
+    const languageOption = bookLang.options[i];
+    if (languageOption.value === "en") {
+      languageOption.selected = detectedLanguage === "www";
+      // break;
+    }
+    if (languageOption.value === detectedLanguage) {
+      console.log(languageOption.value);
+      languageOption.selected = true;
+      break;
+    }
   }
 };
+
+detectBookLang();
 
 // restrict ability to order books for children under 10 month
 function restrictAge() {
@@ -86,193 +86,6 @@ function restrictAge() {
 }
 
 restrictAge();
-// function restrictAge() {
-//   const today = new Date();
-//   const pastYear = new Date(today);
-//   pastYear.setFullYear(today.getFullYear() - 1);
-//   dateInput.setAttribute("max", pastYear.toISOString().split("T")[0]);
-// }
-// restrictAge();
-
-countryInputField.addEventListener("change", (e) => {
-  initCityAutocomplete(e.target.value);
-  autocompleteStr.setComponentRestrictions({
-    country: countryInputField.value,
-  });
-});
-function initCityAutocomplete(selectedCountry = "lv") {
-  autocompleteCity = new google.maps.places.Autocomplete(cityInputField, {
-    types: ["(regions)"],
-  });
-  autocompleteCity.setComponentRestrictions({ country: selectedCountry });
-  autocompleteCity.addListener("place_changed", fillInCity);
-}
-function fillInCity() {
-  const place = autocompleteCity.getPlace();
-  console.log(place);
-  // find address data by type
-  function findAddressData(data) {
-    const dataObject = place.address_components.find((el) =>
-      el.types.includes(data)
-    );
-    return dataObject ? dataObject.long_name : "";
-  }
-
-  if (!place.geometry) {
-    cityInputField.placeholder = "Enter a place";
-  } else {
-    // different address fields for different countries
-    switch (countryInputField.value) {
-      case "lv":
-        cityInputField.value = findAddressData("locality")
-          ? findAddressData("administrative_area_level_1")
-            ? findAddressData("locality") +
-              ", " +
-              findAddressData("administrative_area_level_1")
-            : findAddressData("locality")
-          : findAddressData("administrative_area_level_2") +
-            ", " +
-            findAddressData("administrative_area_level_1");
-        break;
-
-      case "gb":
-        cityInputField.value =
-          findAddressData("locality") +
-          ", " +
-          findAddressData("administrative_area_level_2");
-        break;
-
-      case "de":
-        cityInputField.value =
-          findAddressData("locality") +
-          ", " +
-          findAddressData("administrative_area_level_1");
-        break;
-      default:
-        break;
-    }
-    document.getElementById("ZipCode").value = findAddressData("postal_code");
-  }
-  // rectric autocomplete to city
-  function setCityBias(cityName) {
-    const geocoder = new google.maps.Geocoder();
-    geocoder.geocode({ address: cityName }, (results, status) => {
-      if (status === google.maps.GeocoderStatus.OK) {
-        const cityLocation = results[0].geometry.location;
-        const circle = new google.maps.Circle({
-          center: cityLocation,
-          radius: 10000, // Adjust the radius as needed
-        });
-
-        autocompleteStr.setBounds(circle.getBounds());
-      }
-    });
-  }
-  setCityBias(document.getElementById("City").value);
-}
-// google maps autocomplete for street
-autocompleteStr = new google.maps.places.Autocomplete(
-  document.getElementById("Street"),
-  {
-    types: ["address"],
-  }
-);
-autocompleteStr.addListener("place_changed", fillInAddress);
-
-function fillInAddress() {
-  const address = autocompleteStr.getPlace();
-  console.log(address);
-  function findAddressData(data) {
-    const dataObject = address.address_components.find((el) =>
-      el.types.includes(data)
-    );
-    return dataObject ? dataObject.long_name : "";
-  }
-
-  if (!address.geometry) {
-    streetInputField.placeholder = "Enter a place";
-  } else {
-    // different address fields for different countries
-    switch (countryInputField.value) {
-      case "lv":
-        streetInputField.value = findAddressData("street_number")
-          ? findAddressData("route") + ", " + findAddressData("street_number")
-          : findAddressData("route")
-          ? findAddressData("route")
-          : findAddressData("premise") || findAddressData("establishment");
-        break;
-
-      case "gb":
-        streetInputField.value =
-          findAddressData("route") + ", " + findAddressData("street_number");
-        break;
-
-      case "de":
-        streetInputField.value =
-          findAddressData("route") + ", " + findAddressData("street_number");
-        break;
-      default:
-        break;
-    }
-    document.getElementById("ZipCode").value = findAddressData("postal_code");
-  }
-  validation();
-}
-// validation();
-// // Replace with your API key
-// const apiKey = "AIzaSyCRMOibgbPFCnvJ1IZjVTIVgYPRE7pk0rE";
-
-// // Initialize the Google Places Service
-// const service = new google.maps.places.PlacesService(
-//   document.createElement("div")
-// );
-
-// // Construct the address query
-// const addressQuery = {
-//   query: "YOUR_STREET_NAME, YOUR_CITY, YOUR_COUNTRY",
-//   key: apiKey,
-// };
-// // Perform the search
-// service.textSearch(addressQuery, (results, status) => {
-//   if (status === google.maps.places.PlacesServiceStatus.OK) {
-//     // The results array will contain matching places
-//     if (results.length > 0) {
-//       const firstResult = results[0];
-//       console.log("Found Address:", firstResult.formatted_address);
-//     } else {
-//       console.log("No matching address found.");
-//     }
-//   } else {
-//     console.error("Place search request failed:");
-//   }
-// });
-
-// added dropdown list of countries to phone input
-const phoneInput = window.intlTelInput(phoneInputField, {
-  initialCountry: "auto",
-  geoIpLookup: (callback) => {
-    fetch("https://ipapi.co/json")
-      .then((res) => res.json())
-      .then((data) => callback(data.country_code))
-      .catch(() => callback("us"));
-  },
-  utilsScript:
-    "https://cdnjs.cloudflare.com/ajax/libs/intl-tel-input/17.0.8/js/utils.js",
-});
-
-// added country code to phone input when input is changed
-phoneInputField.addEventListener("change", (e) => {
-  e.target.value = phoneInput.getNumber();
-});
-
-// added new field to form data object which shows if order made during staging or production
-if (host.includes("stastiem.webflow.io") || port !== "") {
-  console.log("staging");
-  environment.value = "staging";
-} else {
-  console.log("production");
-  environment.value = "production";
-}
 
 $(progressbarClone).removeClass("current");
 $('[data-form="progress"]').children().remove();
@@ -285,64 +98,6 @@ steps.hide();
 $('[data-form="next-btn"][type="submit"]').each(function () {
   $(this).attr("type", "button");
 });
-
-function changeImage() {
-  if (image_changed == false) {
-    console.log("Updating 3. step picture, uploading guide");
-
-    const imageElement1 = document.getElementsByClassName("image-5")[0];
-    const imageElement2 = document.getElementById(
-      "w-node-_64c267de-1f7b-60c8-f08e-df7363cd287d-f2b54ff2"
-    );
-    const imageElement3 = document.getElementById(
-      "w-node-_5ab35103-f747-f14b-648c-fa1ee52dda2c-f2b54ff2"
-    );
-    const imageElement4 = document.getElementById(
-      "w-node-c8bf5738-30c7-c414-4e66-bf7f05839cfa-f2b54ff2"
-    );
-    const imageElement5 = document.getElementById(
-      "w-node-_47f06dcd-89d6-8710-a41b-fa0d1a830a72-f2b54ff2"
-    );
-    const imageElement6 = document.getElementById(
-      "w-node-ce4f6a12-c0fa-b8c4-ea5d-29324ea4ee96-f2b54ff2"
-    );
-
-    function updateImage(imageElement, image_name, extension) {
-      url =
-        "https://stastiem-public-assets.s3-accelerate.amazonaws.com/website/" +
-        image_name +
-        extension;
-
-      console.log(
-        "Updating image element: " + imageElement + " with url: " + url + ""
-      );
-      imageElement.src = url;
-      // imageElement.srcset = url + " 500w, " +
-      //                       url + " 800w, " +
-      //                       url + " 1080w, " +
-      //                       url + " 1600w, " +
-      //                       url + " 2000w, ";
-    }
-
-    if (is_boy) {
-      updateImage(imageElement1, "boy1", ".png");
-      updateImage(imageElement2, "boy2", ".png");
-      updateImage(imageElement3, "boy3", ".png");
-      updateImage(imageElement4, "boy4", ".png");
-      updateImage(imageElement5, "boy5", ".png");
-      updateImage(imageElement6, "boy6", ".png");
-    } else {
-      updateImage(imageElement1, "girl1", ".png");
-      updateImage(imageElement2, "girl2", ".png");
-      updateImage(imageElement3, "girl3", ".png");
-      updateImage(imageElement4, "girl4", ".png");
-      updateImage(imageElement5, "girl5", ".png");
-      updateImage(imageElement6, "girl6", ".png");
-    }
-
-    image_changed = true;
-  }
-}
 
 function getParams() {
   urlFormly.searchParams.forEach(function (val, key) {
@@ -417,38 +172,6 @@ if (quiz) {
     $(this).children().attr("data-radio-skip", true);
     $(this).children().attr("data-radio-delay", 250);
   });
-}
-
-function disableBtn() {
-  fill = false;
-  //next button style
-  $('[data-form="next-btn"]').css({
-    opacity: "0.4",
-    "pointer-events": "none",
-  });
-  $('[data-form="next-btn"]').addClass("disabled");
-  //submit btn style
-  $('[data-form="submit-btn"]').css({
-    opacity: "0.4",
-    "pointer-events": "none",
-  });
-  $('[data-form="submit-btn"]').addClass("disabled");
-}
-
-function enableBtn() {
-  fill = true;
-  //next button style
-  $('[data-form="next-btn"]').css({
-    "pointer-events": "auto",
-    opacity: "1",
-  });
-  $('[data-form="next-btn"]').removeClass("disabled");
-  //submit btn style
-  $('[data-form="submit-btn"]').css({
-    "pointer-events": "auto",
-    opacity: "1",
-  });
-  $('[data-form="submit-btn"]').removeClass("disabled");
 }
 
 function saveFilledInput() {
@@ -536,9 +259,9 @@ function updateStep() {
   empReqInput = [];
   empReqSelect = [];
   empReqTextarea = [];
-  //selections = []
+  selections = [];
 
-  //custom clickable progress indicator
+  // custom clickable progress indicator
   if ($("[data-clickable]").data("clickable")) {
     console.log("clickable indicator");
     steps.find(":input[required]").each(function () {
@@ -593,7 +316,7 @@ function updateStep() {
 
   $(progressbar).removeClass("current");
 
-  for (i = 0; i <= x; i++) {
+  for (let i = 0; i <= x; i++) {
     $(progressbar[i]).addClass("current");
   }
   if (reinitIX === true) {
@@ -637,827 +360,39 @@ function updateStep() {
   //focus first input in every step
   $($(steps[x]).find("input[autofocus]")[0]).focus();
   $($(steps[x]).find("textarea[autofocus]")[0]).focus();
-  validation();
+  validation(
+    x,
+    steps,
+    radioFilled,
+    inputFilled,
+    telFilled,
+    dateFilled,
+    selectFilled,
+    emailFilled,
+    fileFilled
+  );
 
-  for (idx = 0; idx <= progress; idx++) {
+  for (let idx = 0; idx <= progress; idx++) {
     $($('[data-form="custom-progress-indicator"]')[idx]).removeClass(
       "disabled"
     );
   }
 }
 
-function validateEmail(email, blockDomain) {
-  let domainEntered = email.includes("@")
-    ? email.split("@")[1].split(".")[0]
-    : [];
-  dom = [];
-  if (blockDomain !== undefined) {
-    blockDomain.split(",").forEach(function (x) {
-      if (x.includes(domainEntered)) {
-        dom.push(domainEntered);
-      }
-    });
-  }
-
-  if (dom.length > 0) {
-    domainAllowed = false;
-  } else {
-    domainAllowed = true;
-  }
-
-  var emailReg = /^([\w-\.]+@([\w-]+\.)+[\w-]{2,20})?$/;
-  if (!emailReg.test(email)) {
-    emailFilled = false;
-  } else {
-    emailFilled = true;
-  }
-}
-
-// function validateEmail(email, blockDomain) {
-//   const isValidEmailFormat = /^([\w-\.]+@([\w-]+\.)+[\w-]{2,20})?$/.test(email);
-//   const domainEntered = email.includes("@")
-//     ? email.split("@")[1].split(".")[0]
-//     : "";
-
-//   let domainAllowed = true;
-//   if (blockDomain !== undefined) {
-//     const blockedDomains = blockDomain.split(",");
-//     domainAllowed = !blockedDomains.some((x) => x.includes(domainEntered));
-//   }
-
-//   return isValidEmailFormat && domainAllowed;
-// }
-
-function validation() {
-  //conditional logic
-
-  if ($(steps[x]).data("card")) {
-    enableBtn();
-  }
-
-  textareaLength = $(steps[x]).find("textarea[required]:visible").length;
-  textInputLength = $(steps[x]).find(
-    'input[type="text"][required]:visible'
-  ).length;
-  selectInputLength = $(steps[x]).find("select[required]:visible").length;
-  emailInputLength = $(steps[x]).find('input[type="email"]:visible').length;
-  checkboxInputLength = $(steps[x]).find(
-    'input[type="checkbox"]:visible'
-  ).length;
-  // console.log($(steps[x]).find('input[type="text"][required]:visible'));
-  if (textInputLength > 0 || selectInputLength > 0 || textareaLength > 0) {
-    disableBtn();
-  } else {
-    enableBtn();
-  }
-
-  var checkCount = $(steps[x]).data("checkbox")
-    ? $(steps[x]).data("checkbox")
-    : 0;
-
-  if (!$("[data-logic-extra]").data("logic-extra")) {
-    if ($(steps[x]).find(":input").is('[type="checkbox"]')) {
-      console.log("checkbox step val: " + checkCount);
-      if (
-        checkCount === "*" ||
-        checkCount > $(steps[x]).find(':input[type="checkbox"]').length
-      ) {
-        $(steps[x])
-          .find(':input[type="checkbox"]')
-          .each(function () {
-            if ($(this).is(":checked")) {
-              if ($(steps[x]).find(":input[required]").length < 1) {
-                checkboxFilled = true;
-              }
-            } else {
-              checkboxFilled = false;
-            }
-          });
-      } else {
-        if (
-          $(steps[x]).find(':input[type="checkbox"]:checked').length >=
-          checkCount
-        ) {
-          checkboxFilled = true;
-        } else {
-          checkboxFilled = false;
-        }
-      }
-    }
-
-    if ($(steps[x]).find(":input[required]").is('[type="radio"]')) {
-      if ($(steps[x]).find(':input[type="radio"]').is(":checked")) {
-        radioFilled = true;
-      } else {
-        radioFilled = false;
-      }
-    }
-
-    $(steps[x])
-      .find(':input[type="text"][required]')
-      .each(function (i) {
-        console.log("Text input step val: " + $(this).val());
-        if ($(this).val() !== "") {
-          empReqInput = empReqInput.filter((y) => y.input !== i);
-        } else {
-          if (!empReqInput.find((y) => y.input === i)) {
-            empReqInput.push({ input: i });
-          }
-        }
-
-        if (empReqInput.length === 0) {
-          inputFilled = true;
-        } else {
-          inputFilled = false;
-        }
-      });
-
-    // const requiredTextInputFields = $(steps[x]).find(
-    //   ':input[type="text"][required]'
-    // );
-
-    // requiredTextInputFields.each(function (i) {
-    //   const inputValue = $(this).val();
-
-    //   console.log(empReqInput);
-    //   console.log("Text input step val: " + inputValue);
-
-    //   const existingInputIndex = empReqInput.findIndex((y) => y.input === i);
-
-    //   if (inputValue !== "") {
-    //     if (existingInputIndex !== -1) {
-    //       empReqInput.splice(existingInputIndex, 1);
-    //     }
-    //   } else {
-    //     if (existingInputIndex === -1) {
-    //       empReqInput.push({ input: i });
-    //     }
-    //   }
-
-    //   inputFilled = empReqInput.length === 0;
-    // });
-
-    $(steps[x])
-      .find(':input[type="tel"][required]')
-      .each(function (i) {
-        console.log("Tel input step val: " + $(this).val());
-
-        if ($(this).val() !== "") {
-          empReqTel = empReqTel.filter((y) => y.input !== i);
-        } else {
-          if (!empReqTel.find((y) => y.input === i)) {
-            empReqTel.push({ input: i });
-          }
-        }
-
-        if (empReqTel.length === 0) {
-          telFilled = true;
-        } else {
-          telFilled = false;
-        }
-      });
-
-    $(steps[x])
-      .find(':input[type="date"][required]')
-      .each(function (i) {
-        console.log("Date input step val: " + $(this).val());
-
-        if ($(this).val() !== "") {
-          empReqDate = empReqDate.filter((y) => y.input !== i);
-        } else {
-          if (!empReqDate.find((y) => y.input === i)) {
-            empReqDate.push({ input: i });
-          }
-        }
-
-        if (empReqDate.length === 0) {
-          dateFilled = true;
-        } else {
-          dateFilled = false;
-        }
-      });
-
-    $(steps[x])
-      .find(':input[type="file"][required]')
-      .each(function (i) {
-        console.log("File input step val: " + $(this).val());
-        changeImage();
-
-        let empReqFile = [];
-
-        if ($(this).val() !== "") {
-          empReqFile = empReqFile.filter((y) => y.input !== i);
-        } else {
-          if (!empReqFile.find((y) => y.input === i)) {
-            empReqFile.push({ input: i });
-          }
-        }
-
-        if (empReqFile.length === 0) {
-          fileFilled = true;
-        } else {
-          fileFilled = false;
-        }
-      });
-
-    // Happens when clicking the "remove" button on the file upload
-    $(steps[x])
-      .find(".w-icon-file-upload-remove")
-      .on("click", function () {
-        console.log("1. User removed a file " + $(this).val());
-        disableBtn();
-      });
-
-    // Happens when clicking the "remove" button on the file upload
-    $(steps[x])
-      .find(".link-5.w-file-remove-link")
-      .on("click", function () {
-        console.log("2. User removed a file " + $(this).val());
-        disableBtn();
-      });
-
-    $(steps[x])
-      .find(".link-back.w-inline-block")
-      .on("click", function () {
-        console.log("User clicked back" + $(this).val());
-        enableBtn();
-        setAllChecksToTrue();
-      });
-
-    $(steps[x])
-      .find("select[required]")
-      .each(function (i) {
-        console.log("Select input step val: " + $(this).val());
-
-        if ($(this).val() !== "") {
-          empReqSelect = empReqSelect.filter((y) => y.input !== i);
-        } else {
-          if (!empReqSelect.find((y) => y.input === i)) {
-            empReqSelect.push({ input: i });
-          }
-        }
-
-        if (empReqSelect.length === 0) {
-          selectFilled = true;
-        } else {
-          selectFilled = false;
-        }
-      });
-
-    $(steps[x])
-      .find("textarea[required]")
-      .each(function (i) {
-        console.log("Textarea input step val: " + $(this).val());
-        if ($(this).val() !== "") {
-          empReqTextarea = empReqTextarea.filter((y) => y.input !== i);
-        } else {
-          if (!empReqTextarea.find((y) => y.input === i)) {
-            empReqTextarea.push({ input: i });
-          }
-        }
-
-        if (empReqTextarea.length === 0) {
-          textareaFilled = true;
-        } else {
-          textareaFilled = false;
-        }
-      });
-
-    $(steps[x])
-      .find(':input[type="email"][required]')
-      .each(function () {
-        console.log("Email input step val: " + $(this).val());
-        if ($(this).val() !== "") {
-          validateEmail($(this).val(), $(this).data("block-domain"));
-        } else {
-          emailFilled = false;
-        }
-      });
-  } else {
-    //////////////////////////////////logic extra validation//////////////////////////////////////////////////
-    if ($(steps[x]).data("card")) {
-      answer = $(steps[x]).find("[data-go-to]").data("go-to");
-      selections = selections.filter((y) => y.step !== x);
-      selections.push({ step: x, selected: answer });
-      console.log(selections);
-    }
-
-    /////////////////////////////////checkbox validation//////////////////////////////////////
-    // if (
-    //   $(steps[x])
-    //     .find("[data-answer]:visible")
-    //     .find(":input")
-    //     .is('[type="checkbox"]')
-    // ) {
-    //   if (
-    //     checkCount === "*" ||
-    //     checkCount > $(steps[x]).find(':input[type="checkbox"]').length
-    //   ) {
-    //     $(steps[x])
-    //       .find(':input[type="checkbox"]')
-    //       .each(function () {
-    //         console.log("Checkbox input step val: " + $(this).val());
-    //         if ($(this).is(":checked")) {
-    //           if ($(steps[x]).find(":input[required]").length < 1) {
-    //             if ($(this).parents("[data-go-to]").attr("data-go-to")) {
-    //               answer = $(this).parents("[data-go-to]").attr("data-go-to");
-    //               selections = selections.filter((y) => y.step !== x);
-    //               selections.push({ step: x, selected: answer });
-    //             }
-    //             checkboxFilled = true;
-    //           }
-    //         } else {
-    //           checkboxFilled = false;
-    //         }
-    //       });
-    //   } else {
-    //     if (
-    //       $(steps[x])
-    //         .find("[data-answer]:visible")
-    //         .find(':input[type="checkbox"]:checked').length >= checkCount
-    //     ) {
-    //       console.log("Checkbox input step val: " + $(this).val());
-
-    //       if (
-    //         $(steps[x])
-    //           .find("[data-answer]:visible")
-    //           .find(':input[type="checkbox"]:checked')
-    //           .parents("[data-go-to]")
-    //           .attr("data-go-to")
-    //       ) {
-    //         answer = $(steps[x])
-    //           .find("[data-answer]:visible")
-    //           .find(':input[type="checkbox"]:checked')
-    //           .parents("[data-go-to]")
-    //           .attr("data-go-to");
-    //         selections = selections.filter((y) => y.step !== x);
-    //         selections.push({ step: x, selected: answer });
-    //       }
-    //       selections = selections.filter((y) => y.step !== x);
-    //       selections.push({ step: x, selected: answer });
-    //       checkboxFilled = true;
-    //       //}
-    //     } else {
-    //       checkboxFilled = false;
-    //     }
-    //   }
-    // }
-
-    //////////////////////////////////radio input validation////////////////////////////////////////////
-    // if (
-    //   $(steps[x])
-    //     .find("[data-answer]:visible")
-    //     .find(":input[required]")
-    //     .is('[type="radio"]')
-    // ) {
-    //   if (
-    //     $(steps[x])
-    //       .find("[data-answer]:visible")
-    //       .find(':input[type="radio"][required]')
-    //       .is(":checked")
-    //   ) {
-    //     console.log("Radio input step val: " + $(this).val());
-    //     radioFilled = true;
-    //   } else {
-    //     radioFilled = false;
-    //   }
-    // } else {
-    //   radioFilled = true;
-    // }
-
-    //////////////////////////text input validation/////////////////////////////////////
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find(':input[type="text"][required]')
-    //   .each(function (i) {
-    //     console.log("Text input step val: " + $(this).val());
-    //     if ($(this).val() !== "") {
-    //       empReqInput = empReqInput.filter((y) => y.input !== i);
-    //     } else {
-    //       if (!empReqInput.find((y) => y.input === i)) {
-    //         empReqInput.push({ input: i });
-    //       }
-    //     }
-    //     console.log($(steps[x]).find("[data-answer]:visible"));
-    //     if (empReqInput.length === 0) {
-    //       inputFilled = true;
-    //     } else {
-    //       inputFilled = false;
-    //     }
-    //   });
-
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find(':input[type="text"]')
-    //   .each(function (i) {
-    //     console.log("Text input step val: " + $(this).val());
-    //     skipTo = undefined;
-    //     if ($(this).parents("[data-skip-to]").data("skip-to") !== "") {
-    //       skipTo = $(this).parents("[data-skip-to]").data("skip-to");
-    //     }
-    //     if ($(this).parents("[data-go-to]").attr("data-go-to")) {
-    //       answer = $(this).parents("[data-go-to]").attr("data-go-to");
-    //       selections = selections.filter((y) => y.step !== x);
-    //       selections.push({ step: x, selected: answer });
-    //       if (skipTo) {
-    //         selections.push({ step: skipTo - 2, selected: answer });
-    //         objIndex = selections.findIndex((obj) => obj.step === x);
-    //         selections[objIndex].skipTo = parseInt(skipTo) - 1;
-    //         selections[objIndex].backTo = x;
-    //       }
-    //     }
-    //   });
-
-    //////////////////////////phone input validation/////////////////////////////////////
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find(':input[type="tel"][required]')
-    //   .each(function (i) {
-    //     console.log("Phone input step val: " + $(this).val());
-    //     if ($(this).val() !== "") {
-    //       empReqTel = empReqTel.filter((y) => y.input !== i);
-    //     } else {
-    //       if (!empReqTel.find((y) => y.input === i)) {
-    //         empReqTel.push({ input: i });
-    //       }
-    //     }
-
-    //     if (empReqTel.length === 0) {
-    //       telFilled = true;
-    //     } else {
-    //       telFilled = false;
-    //     }
-    //   });
-
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find(':input[type="tel"]')
-    //   .each(function (i) {
-    //     console.log("Phone input step val: " + $(this).val());
-    //     skipTo = undefined;
-    //     if ($(this).parents("[data-skip-to]").data("skip-to") !== "") {
-    //       skipTo = $(this).parents("[data-skip-to]").data("skip-to");
-    //     }
-    //     if ($(this).parents("[data-go-to]").attr("data-go-to")) {
-    //       answer = $(this).parents("[data-go-to]").attr("data-go-to");
-    //       selections = selections.filter((y) => y.step !== x);
-    //       selections.push({ step: x, selected: answer });
-    //       if (skipTo) {
-    //         selections.push({ step: skipTo - 2, selected: answer });
-    //         objIndex = selections.findIndex((obj) => obj.step === x);
-    //         selections[objIndex].skipTo = parseInt(skipTo) - 1;
-    //         selections[objIndex].backTo = x;
-    //       }
-    //     }
-    //   });
-
-    //////////////////////////file input validation/////////////////////////////////////
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find(':input[type="file"][required]')
-    //   .each(function (i) {
-    //     console.log("File input validation val: " + $(this).val());
-
-    //     if ($(this).val() !== "") {
-    //       empReqfile = empReqfile.filter((y) => y.input !== i);
-    //     } else {
-    //       if (!empReqfile.find((y) => y.input === i)) {
-    //         empReqfile.push({ input: i });
-    //       }
-    //     }
-
-    //     if (empReqfile.length === 0) {
-    //       fileFilled = true;
-    //     } else {
-    //       fileFilled = false;
-    //     }
-    //   });
-
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find(':input[type="file"]')
-    //   .each(function (i) {
-    //     console.log("File input validation val: " + $(this).val());
-
-    //     skipTo = undefined;
-    //     if ($(this).parents("[data-skip-to]").data("skip-to") !== "") {
-    //       skipTo = $(this).parents("[data-skip-to]").data("skip-to");
-    //     }
-    //     if ($(this).parents("[data-go-to]").attr("data-go-to")) {
-    //       answer = $(this).parents("[data-go-to]").attr("data-go-to");
-    //       selections = selections.filter((y) => y.step !== x);
-    //       selections.push({ step: x, selected: answer });
-    //       if (skipTo) {
-    //         selections.push({ step: skipTo - 2, selected: answer });
-    //         objIndex = selections.findIndex((obj) => obj.step === x);
-    //         selections[objIndex].skipTo = parseInt(skipTo) - 1;
-    //         selections[objIndex].backTo = x;
-    //       }
-    //     }
-    //   });
-
-    //////////////////////////number input validation/////////////////////////////////////
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find(':input[type="number"][required]')
-    //   .each(function (i) {
-    //     console.log("Number input step val: " + $(this).val());
-    //     if ($(this).val() !== "") {
-    //       empReqInput = empReqInput.filter((y) => y.input !== i);
-    //     } else {
-    //       if (!empReqInput.find((y) => y.input === i)) {
-    //         empReqInput.push({ input: i });
-    //       }
-    //     }
-
-    //     if (empReqInput.length === 0) {
-    //       inputFilled = true;
-    //     } else {
-    //       inputFilled = false;
-    //     }
-    //   });
-
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find(':input[type="number"]')
-    //   .each(function (i) {
-    //     console.log("Number input step val: " + $(this).val());
-    //     skipTo = undefined;
-    //     if ($(this).parents("[data-skip-to]").data("skip-to") !== "") {
-    //       skipTo = $(this).parents("[data-skip-to]").data("skip-to");
-    //     }
-    //     if ($(this).parents("[data-go-to]").attr("data-go-to")) {
-    //       answer = $(this).parents("[data-go-to]").attr("data-go-to");
-    //       selections = selections.filter((y) => y.step !== x);
-    //       selections.push({ step: x, selected: answer });
-    //       if (skipTo) {
-    //         selections.push({ step: skipTo - 2, selected: answer });
-    //         objIndex = selections.findIndex((obj) => obj.step === x);
-    //         selections[objIndex].skipTo = parseInt(skipTo) - 1;
-    //         selections[objIndex].backTo = x;
-    //       }
-    //     }
-    //   });
-
-    //////////////////////////select input validation///////////////////////////////////
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find("select[required]")
-    //   .each(function (i) {
-    //     console.log("Select input step val: " + $(this).val());
-    //     if ($(this).val() !== "") {
-    //       empReqSelect = empReqSelect.filter((y) => y.input !== i);
-    //     } else {
-    //       if (!empReqSelect.find((y) => y.input === i)) {
-    //         empReqSelect.push({ input: i });
-    //       }
-    //     }
-
-    //     if (empReqSelect.length === 0) {
-    //       selectFilled = true;
-    //     } else {
-    //       selectFilled = false;
-    //     }
-    //   });
-
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find("select")
-    //   .each(function (i) {
-    //     console.log("Select input step val: " + $(this).val());
-    //     skipTo = undefined;
-    //     if ($(this).parents("[data-skip-to]").data("skip-to") !== "") {
-    //       skipTo = $(this).parents("[data-skip-to]").data("skip-to");
-    //     }
-    //     if ($(this).parents("[data-go-to]").attr("data-go-to")) {
-    //       answer = $(this).parents("[data-go-to]").attr("data-go-to");
-    //       selections = selections.filter((y) => y.step !== x);
-    //       selections.push({ step: x, selected: answer });
-    //       if (skipTo) {
-    //         selections.push({ step: skipTo - 2, selected: answer });
-    //         objIndex = selections.findIndex((obj) => obj.step === x);
-    //         selections[objIndex].skipTo = parseInt(skipTo) - 1;
-    //         selections[objIndex].backTo = x;
-    //       }
-    //     }
-    //   });
-
-    //////////////////////////textarea validation////////////////////////////////
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find("textarea[required]")
-    //   .each(function (i) {
-    //     console.log("Textarea input step val: " + $(this).val());
-    //     if ($(this).val() !== "") {
-    //       empReqTextarea = empReqTextarea.filter((y) => y.input !== i);
-    //     } else {
-    //       if (!empReqTextarea.find((y) => y.input === i)) {
-    //         empReqTextarea.push({ input: i });
-    //       }
-    //     }
-
-    //     if (empReqTextarea.length === 0) {
-    //       textareaFilled = true;
-    //     } else {
-    //       textareaFilled = false;
-    //     }
-    //   });
-
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find("textarea")
-    //   .each(function (i) {
-    //     console.log("Textarea input step val: " + $(this).val());
-    //     skipTo = undefined;
-    //     if ($(this).parents("[data-skip-to]").data("skip-to") !== "") {
-    //       skipTo = $(this).parents("[data-skip-to]").data("skip-to");
-    //     }
-    //     if ($(this).parents("[data-go-to]").attr("data-go-to")) {
-    //       answer = $(this).parents("[data-go-to]").attr("data-go-to");
-    //       selections = selections.filter((y) => y.step !== x);
-    //       selections.push({ step: x, selected: answer });
-    //       if (skipTo) {
-    //         selections.push({ step: skipTo - 2, selected: answer });
-    //         objIndex = selections.findIndex((obj) => obj.step === x);
-    //         selections[objIndex].skipTo = parseInt(skipTo) - 1;
-    //         selections[objIndex].backTo = x;
-    //       }
-    //     }
-    //   });
-
-    /////////////////////////email validation//////////////////////////////////////
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find(':input[type="email"][required]')
-    //   .each(function (m) {
-    //     console.log("Email input step val: " + $(this).val());
-    //     if ($(this).val() !== "") {
-    //       validateEmail($(this).val(), $(this).data("block-domain"));
-    //     } else {
-    //       emailFilled = false;
-    //     }
-    //   });
-
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find(':input[type="email"]')
-    //   .each(function (m) {
-    //     console.log("Email input step val: " + $(this).val());
-    //     skipTo = undefined;
-    //     if ($(this).parents("[data-skip-to]").data("skip-to") !== "") {
-    //       skipTo = $(this).parents("[data-skip-to]").data("skip-to");
-    //     }
-    //     if ($(this).parents("[data-go-to]").attr("data-go-to")) {
-    //       answer = $(this).parents("[data-go-to]").attr("data-go-to");
-    //       selections = selections.filter((y) => y.step !== x);
-    //       selections.push({ step: x, selected: answer });
-    //       if (skipTo) {
-    //         selections.push({ step: skipTo - 2, selected: answer });
-    //         objIndex = selections.findIndex((obj) => obj.step === x);
-    //         selections[objIndex].skipTo = parseInt(skipTo) - 1;
-    //         selections[objIndex].backTo = x;
-    //       }
-    //     }
-    //   });
-
-    /////////////////////////date validation//////////////////////////////////////
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find(':input[type="date"][required]')
-    //   .each(function (m) {
-    //     console.log("date", $(this).val());
-
-    //     if ($(this).val() !== "") {
-    //       empReqDate = empReqDate.filter((y) => y.input !== m);
-    //     } else {
-    //       if (!empReqDate.find((y) => y.input === m)) {
-    //         empReqDate.push({ input: m });
-    //       }
-    //     }
-
-    //     if (empReqDate.length === 0) {
-    //       dateFilled = true;
-    //     } else {
-    //       dateFilled = false;
-    //     }
-    //   });
-
-    // $(steps[x])
-    //   .find("[data-answer]:visible")
-    //   .find(':input[type="date"]')
-    //   .each(function (m) {
-    //     console.log("date2", $(this).val());
-    //     skipTo = undefined;
-    //     if ($(this).parents("[data-skip-to]").data("skip-to") !== "") {
-    //       skipTo = $(this).parents("[data-skip-to]").data("skip-to");
-    //     }
-    //     if ($(this).parents("[data-go-to]").attr("data-go-to")) {
-    //       answer = $(this).parents("[data-go-to]").attr("data-go-to");
-    //       selections = selections.filter((y) => y.step !== x);
-    //       selections.push({ step: x, selected: answer });
-    //       if (skipTo) {
-    //         selections.push({ step: skipTo - 2, selected: answer });
-    //         objIndex = selections.findIndex((obj) => obj.step === x);
-    //         selections[objIndex].skipTo = parseInt(skipTo) - 1;
-    //         selections[objIndex].backTo = x;
-    //       }
-    //     }
-    //   });
-  }
-
-  function setAllChecksToTrue() {
-    inputFilled = true;
-    dateFilled = true;
-    checkboxFilled = true;
-    telFilled = true;
-    radioFilled = true;
-    emailFilled = true;
-    domainAllowed = true;
-    selectFilled = true;
-    fileFilled = true;
-    textareaFilled = true;
-  }
-
-  if (
-    inputFilled &&
-    dateFilled &&
-    checkboxFilled &&
-    telFilled &&
-    radioFilled &&
-    emailFilled &&
-    domainAllowed &&
-    selectFilled &&
-    fileFilled &&
-    textareaFilled
-  ) {
-    console.log(
-      "ENABLING BOTTON BECAUSE inputFilled: " +
-        inputFilled +
-        " dateFilled: " +
-        dateFilled +
-        "checkboxFilled: " +
-        checkboxFilled +
-        " telFilled: " +
-        telFilled +
-        " radioFilled: " +
-        radioFilled +
-        " emailFilled: " +
-        emailFilled +
-        " domainAllowed: " +
-        domainAllowed +
-        " selectFilled: " +
-        selectFilled +
-        " fileFilled: " +
-        fileFilled +
-        " textareaFilled: " +
-        textareaFilled
-    );
-    enableBtn();
-  } else {
-    console.log(
-      "DISABLING BOTTON BECAUSE inputFilled: " +
-        inputFilled +
-        " dateFilled: " +
-        dateFilled +
-        "checkboxFilled: " +
-        checkboxFilled +
-        " telFilled: " +
-        telFilled +
-        " radioFilled: " +
-        radioFilled +
-        " emailFilled: " +
-        emailFilled +
-        " domainAllowed: " +
-        domainAllowed +
-        " selectFilled: " +
-        selectFilled +
-        " fileFilled: " +
-        fileFilled +
-        " textareaFilled: " +
-        textareaFilled
-    );
-    disableBtn();
-  }
-}
-
 function nextStep() {
   x++;
-  console.log(memory);
   if (x > progress) {
     progress = x;
   }
-
   if (x <= steps.length - 1) {
     updateStep();
     if (memory) {
       saveFilledInput();
     }
-
     $('[data-text="current-step"]').text(
       $(steps[x]).data("card") ? (curStep = x + 0) : (curStep = x + 1)
     );
   }
-  console.log(x);
   return x;
 }
 
@@ -1528,7 +463,17 @@ $('[data-form="back-btn"]').on("click", function () {
 $(steps)
   .find(":input")
   .on("input", function (input) {
-    validation();
+    validation(
+      x,
+      steps,
+      radioFilled,
+      inputFilled,
+      telFilled,
+      dateFilled,
+      selectFilled,
+      emailFilled,
+      fileFilled
+    );
   });
 
 $(steps)
@@ -1558,7 +503,7 @@ $(steps)
       } else {
         is_boy = true;
       }
-
+      changeImage(image_changed, is_boy);
       $(steps[x])
         .find("[data-answer]:visible")
         .find(":input[type='radio']:checked")
@@ -1613,6 +558,7 @@ if ($("[data-clickable-all]").data("clickable-all")) {
   $('[data-form="custom-progress-indicator"]').addClass("disabled");
 }
 function clickableIndicator() {
+  console.log("qweqweqwe");
   $('[data-form="progress-indicator"]').removeClass("current");
   if ($("[data-clickable]").data("clickable")) {
     if ($("[data-clickable]").data("clickable-all")) {
@@ -1699,7 +645,3 @@ $("textarea").keypress(function (event) {
     $(this).val($(this).val() + "\n");
   }
 });
-
-// $('[data-form="multistep"]').on("submit", function () {
-//   // here you can return false, to avoid redirect
-// });
